pyproject.toml
[project]
name = "update-checker"
version = "0.1.0"
description = "Checks for outdated packages"
readme = {file = "README.md", content-type = "text/markdown"}
license = "MIT"
authors = [
    {name = "Blake Riley", email = "blake-riley@users.noreply.github.com"},
]
requires-python = ">=3.14,<4.0"
dependencies = [
    "pydantic>=1.8,<2",
    "pyyaml>=6.0,<7",
    "requests>=2.31,<3",
]

[project.scripts]
update-check = "update_checker.update_check:_entrypoint"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=24.3.0,<25",
    "jupyterlab>=3.2.8,<4",
    "mypy>=1.0,<2",
    "pycodestyle>=2.8,<3",
    "pydocstyle>=6.1,<7",
    "pytest>=7.2,<8",
    "types-pyyaml>=6.0,<7",
    "types-requests>=2.31,<3",
]
